+++ src/AtomUI.Desktop.Controls/WindowTitleBar/WindowTitleBarTitleAlignment.cs
@@ -1,0 +1,10 @@
+namespace AtomUI.Desktop.Controls;
+
+public enum WindowTitleBarTitleAlignment
+{
+    Auto,
+    Left,
+    Center,
+    WindowCenter,
+    Right
+}